Extract fpl types into it's own file
